Introduct `jest` for testing components
This patch introduces `jest` as a test suite, which provides a nice was
of testing React components with minimal configuration. `jest` provides
support for _snapshot_ testing which we can use to catch visual
regressions in components.

As of this commit there is an issue with `enzyme` not being able to find
dependencies. Instead of adding those dependencies to the project (which
are not needed otherwise) I have worked around this by entering the
`node_modules/enzyme` directory and running `npm install`. Before this
merges into **master** we should have this issue resolved.